+++ src/factorial.h
@@ -5,8 +5,8 @@
 #ifndef __RAMANUJANFACTORIAL__
 #define __RAMANUJANFACTORIAL__
 
-inline double _ramanujan(unsigned int n) {
-  double f = sqrt(M_PI) * pow(n, n) * exp(-n) * pow((8 * pow(n, 3) + 4 * pow(n, 2) + n + 1 / 30), 1 / 6);
+inline long double _ramanujan(unsigned int n) {
+  long double f= sqrt(M_PI) * pow(n, n) * exp(-n) * pow((8 * pow(n, 3) + 4 * pow(n, 2) + n + 1 / 30), 1 / 6);
   return f;
 }
 
@@ -16,8 +16,8 @@ inline double _ramanujan(unsigned int n) {
 #ifndef __STIRLINGFACTORIAL__
 #define __STIRLINGFACTORIAL__
 
-inline double _stirling(unsigned int n) {
-  double f = sqrt((2.0 * n + 1.0 / 3.0) * M_PI) * pow(n, n) * exp(-n);
+inline long double _stirling(unsigned int n) {
+  long double f = sqrt((2.0 * n + 1.0 / 3.0) * M_PI) * pow(n, n) * exp(-n);
 
   return f;
 }
@@ -28,8 +28,8 @@ inline double _stirling(unsigned int n) {
 #ifndef __STIRLINGLNFACTORIAL__
 #define __STIRLINGLNFACTORIAL__
 
-inline double _stirlingln(unsigned int n) {
-  double f = n * log(n) - n + (1.0 / 6.0) * log(n * (1.0 + 4.0 * n * (1.0 + 2.0 * n))) + 0.5 * log(M_PI);
+inline long double _stirlingln(unsigned int n) {
+  long double f = n * log(n) - n + (1.0 / 6.0) * log(n * (1.0 + 4.0 * n * (1.0 + 2.0 * n))) + 0.5 * log(M_PI);
 
   return f;
 }
